7_4.py
Pizza Toppings: Write a loop that prompts the user to enter a series of pizza toppings until they enter a 'quit' value. As they enter each topping, print a message saying you’ll add that topping to their pizza.

Use a conditional test in the while statement to stop the loop.
Use an active variable to control how long the loop runs.
When complete, print a formatted message that lists the pizza toppings.

Look at the parrot_flag.py program for an example of an active variable.

Please note: parrot_flag.py program was created by my professor and is unavailable for public viewing.