7_10.py
Dream Vacation: Write a program that polls users about their dream vacation. Write a prompt that first asks for the person's name, then asks what country their first dream vacation would be in and then asks what country their second dream vacation would be in. Store the result in a dictionary with the name as the key and the value is a list of destinations. Include a block of code that prints the results of the poll.  See mountain_poll.py